Move files and add screens components
