+++ src/manifest.json
@@ -15,6 +15,10 @@
       "38": "default_icon@2x.png"
     }
   },
+  "options_ui": {
+    "chrome_style": true,
+    "page": "popup.html#settings"
+  },
   "icons": {
     "36": "extension_icon_36.png",
     "128": "extension_icon_128.png"
