+++ src/popup/app.css
@@ -58,11 +58,9 @@
   }
 }
 
-.app {
+body {
   background-color: var(--global-background);
   color: var(--global-color);
-  min-width: 250px;
-  max-width: 320px;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
   font-size: 13px;
 }
