+++ src/popup/screens/popup/popup.css
@@ -1,0 +1,4 @@
+.popup {
+  min-width: 250px;
+  max-width: 320px;
+}
