+++ src/popup/screens/settings/settings.css
@@ -1,0 +1,5 @@
+.settings {
+  min-width: 250px;
+  max-width: 320px;
+  padding: 10px 20px 20px;
+}
